python/duckdb/requirements.txt
pandas
pyarrow
flytekit
flytekitplugins-duckdb
duckdb
